Stacks are Delegates too
